app/backend/requirements.txt
# TERA Backend Requirements
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.25
geoalchemy2==0.14.3
psycopg2-binary==2.9.9

# Task Queue
celery[redis]==5.3.6
redis==5.0.1

# Data Processing
pandas==2.2.0
geopandas==0.14.2
shapely==2.0.2
h3==3.7.6
pyproj==3.6.1

# LLM & AI
httpx>=0.25.2,<0.26.0
ollama==0.1.6
chromadb==0.4.22

# Validation & Config
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Web Scraping
aiohttp==3.9.1
beautifulsoup4==4.12.3
lxml==5.1.0

# Utilities
rich==13.7.0
loguru==0.7.2
tenacity==8.2.3
orjson==3.9.12
geojson==3.1.0
Pillow>=10.0.0
global-land-mask>=0.0.6